+++ Front-end/package.json
@@ -42,6 +42,7 @@
     "@ant-design/icons": "^4.2.2",
     "antd": "^4.5.4",
     "axios": "^0.19.2",
+    "compressorjs": "^1.0.6",
     "formik": "^2.1.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
