+++ Front-end/src/commons/utils/layout.scss
@@ -53,6 +53,9 @@
     &-between {
       justify-content: space-between;
     }
+    &-end {
+      justify-content: flex-end;
+    }
   }
 }
