+++ Front-end/src/containers/AdminPage/ProductPage/ProductAddForm/Camera/index.js
