+++ Front-end/src/containers/AdminPage/ProductPage/ProductAddForm/Mouse/index.js
